Homework is done

+++ src/main/java/mate/academy/springboot/datajpa/repository/ProductRepository.java
@@ -1,5 +1,6 @@
 package mate.academy.springboot.datajpa.repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 import mate.academy.springboot.datajpa.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -9,5 +10,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface ProductRepository extends JpaRepository<Product, Long>,
         JpaSpecificationExecutor<Product> {
-    List<Product> findAllByPriceBetween(double from, double to);
+    List<Product> findAllByPriceBetween(BigDecimal from, BigDecimal to);
 }
