+++ src/main/java/mate/academy/springboot/datajpa/service/ProductService.java
@@ -1,5 +1,6 @@
 package mate.academy.springboot.datajpa.service;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 import mate.academy.springboot.datajpa.model.Product;
@@ -11,7 +12,7 @@ public interface ProductService {
 
     void delete(Long id);
 
-    List<Product> getAllBuThePriceBetween(Double from, Double to);
+    List<Product> getAllBuThePriceBetween(BigDecimal from, BigDecimal to);
 
     List<Product> getAllByTheCategoriesIn(Map<String, String> params);
 }
