+++ src/main/java/mate/academy/springboot/datajpa/service/impl/ProductServiceImpl.java
@@ -1,5 +1,6 @@
 package mate.academy.springboot.datajpa.service.impl;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 import mate.academy.springboot.datajpa.model.Product;
@@ -36,7 +37,7 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public List<Product> getAllBuThePriceBetween(Double from, Double to) {
+    public List<Product> getAllBuThePriceBetween(BigDecimal from, BigDecimal to) {
         return productRepository.findAllByPriceBetween(from, to);
     }
